requirements.txt
Django==1.7
django-autoslug==1.7.2
django-cities-light==3.1.0
six==1.9.0
Unidecode==0.4.17
